refactor locations of behaviors, forces, bases
