+++ docs/reference/base.md
@@ -1,8 +1,13 @@
-# Base Classes
+# Bases
 
 These classes are not generally needed, but are included for completeness.
 
 ```{eval-rst}
 .. currentmodule:: ipyforcegraph
 .. automodule:: ipyforcegraph._base
+```
+
+```{eval-rst}
+.. currentmodule:: ipyforcegraph
+.. automodule:: ipyforcegraph.behaviors._base
 ```
