+++ docs/reference/behaviors.md
@@ -1,6 +1,22 @@
 # Behaviors
 
+## Graph Behaviors
+
 ```{eval-rst}
 .. currentmodule:: ipyforcegraph
-.. automodule:: ipyforcegraph.behaviors
+.. automodule:: ipyforcegraph.behaviors.graph
+```
+
+## Link Behaviors
+
+```{eval-rst}
+.. currentmodule:: ipyforcegraph
+.. automodule:: ipyforcegraph.behaviors.link
+```
+
+## Node Behaviors
+
+```{eval-rst}
+.. currentmodule:: ipyforcegraph
+.. automodule:: ipyforcegraph.behaviors.node
 ```
