+++ src/ipyforcegraph/behaviors/__init__.py
@@ -1,0 +1,41 @@
+# Copyright (c) 2023 ipyforcegraph contributors.
+# Distributed under the terms of the Modified BSD License.
+
+from ._base import Behavior
+from .graph import GraphData, GraphForces, GraphImage
+from .link import (
+    LinkColors,
+    LinkDirectionalArrowColor,
+    LinkDirectionalArrowLength,
+    LinkDirectionalArrowRelPos,
+    LinkDirectionalParticleColor,
+    LinkDirectionalParticles,
+    LinkDirectionalParticleSpeed,
+    LinkDirectionalParticleWidth,
+    LinkLabels,
+    LinkSelection,
+    LinkWidths,
+)
+from .node import NodeColors, NodeLabels, NodeSelection, NodeSizes
+
+__all__ = [
+    "Behavior",
+    "GraphData",
+    "GraphForces",
+    "GraphImage",
+    "LinkColors",
+    "LinkDirectionalArrowColor",
+    "LinkDirectionalArrowLength",
+    "LinkDirectionalArrowRelPos",
+    "LinkDirectionalParticleColor",
+    "LinkDirectionalParticles",
+    "LinkDirectionalParticleSpeed",
+    "LinkDirectionalParticleWidth",
+    "LinkLabels",
+    "LinkSelection",
+    "LinkWidths",
+    "NodeColors",
+    "NodeLabels",
+    "NodeSelection",
+    "NodeSizes",
+]
