+++ src/ipyforcegraph/behaviors/link.py
@@ -1,9 +1,4 @@
-"""Behaviors for ``ipyforcegraph``.
-
-Some special behaviors are more complicated, and have their own modules, such as:
-
-- :class:`~ipyforcegraph.forces.GraphForcesBehavior`
-"""
+"""Link behaviors for ``ipyforcegraph``."""
 
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
@@ -14,159 +9,6 @@ import ipywidgets as W
 import traitlets as T
 
 from ._base import Behavior
-from .sources.dataframe import DataFrameSource
-
-
-@W.register
-class GraphImage(Behavior):
-    """Captures multiple subsequent frames of a canvas, each as an :class:`~ipywidgets.widgets.widget_media.Image`."""
-
-    _model_name: str = T.Unicode("GraphImageModel").tag(sync=True)
-
-    capturing = T.Bool(False, help="Whether the frame capture is currently active").tag(
-        sync=True
-    )
-
-    frame_count = T.Int(1, help="The number of frames to capture").tag(sync=True)
-
-    frames: Tuple[W.Image, ...] = W.TypedTuple(
-        T.Instance(W.Image),
-        help="A tuple of :class:`~ipywidgets.widgets.widget_media.Image` to populate with frames of the graph.",
-    ).tag(sync=True, **W.widget_serialization)
-
-    def _get_frames(self) -> Tuple[W.Image, ...]:
-        return tuple(
-            [W.Image(description=f"frame {i}") for i in range(self.frame_count)]
-        )
-
-    @T.default("frames")
-    def _default_frames(self) -> Tuple[W.Image, ...]:
-        return self._get_frames()
-
-    @T.observe("frame_count")
-    def _on_frame_count(self, change: T.Bunch) -> None:
-        frames = self.frames
-
-        self.frames = tuple()
-
-        for frame in frames:
-            frame.close()
-
-        self.frames = self._get_frames()
-
-
-@W.register
-class GraphData(Behavior):
-    """Captures multiple subsequent ticks of a graph simulation, each as a :class:`~pandas.DataFrame`."""
-
-    _model_name: str = T.Unicode("GraphDataModel").tag(sync=True)
-
-    capturing: bool = T.Bool(
-        False, help="Whether the dataframe capture is currently active"
-    ).tag(sync=True)
-
-    source_count = T.Int(1, help="The number of sources to capture").tag(sync=True)
-
-    sources: Tuple[DataFrameSource, ...] = W.TypedTuple(
-        T.Instance(DataFrameSource),
-        help="A tuple of :class:`~ipyforcegraph.sources.dataframe.DataFrameSource` to be populated with data of the graph.",
-    ).tag(sync=True, **W.widget_serialization)
-
-    def _get_sources(self) -> Tuple[DataFrameSource, ...]:
-        return tuple([DataFrameSource() for i in range(self.source_count)])
-
-    @T.default("sources")
-    def _default_sources(self) -> Tuple[DataFrameSource, ...]:
-        return self._get_sources()
-
-    @T.observe("source_count")
-    def _on_source_count(self, change: T.Bunch) -> None:
-        sources = self.sources
-
-        self.sources = tuple()
-
-        for source in sources:
-            source.close()
-
-        self.sources = self._get_sources()
-
-
-@W.register
-class NodeSelection(Behavior):
-    """Enable node selection with synced ids of selected nodes."""
-
-    _model_name: str = T.Unicode("NodeSelectionModel").tag(sync=True)
-
-    selected: Tuple[Union[int, str], ...] = W.TypedTuple(
-        T.Union((T.Int(), T.Unicode())),
-        allow_none=True,
-        help="the ids of any selected nodes",
-    ).tag(sync=True)
-
-    multiple: bool = T.Bool(True).tag(sync=True)
-
-    selected_color: str = T.Unicode(
-        "rgba(179, 163, 105, 1.0)",
-        help="the color of selected nodes",
-    ).tag(sync=True)
-
-
-@W.register
-class NodeLabels(Behavior):
-    """Display node labels."""
-
-    _model_name: str = T.Unicode("NodeLabelModel").tag(sync=True)
-
-    column_name: str = T.Unicode(
-        None,
-        allow_none=True,
-        help=(
-            "name of the source column to use for node labels. If `None`, use "
-            "the source's `node_id_column`."
-        ),
-    ).tag(sync=True)
-
-    template: Optional[str] = T.Unicode(
-        None,
-        allow_none=True,
-        help="a nunjucks template to use to calculate node labels",
-    ).tag(sync=True)
-
-
-@W.register
-class NodeSizes(Behavior):
-    """Change the size of nodes."""
-
-    _model_name: str = T.Unicode("NodeSizeModel").tag(sync=True)
-
-    column_name: str = T.Unicode(
-        None,
-        allow_none=True,
-        help="name of the source column to use for node sizes.",
-    ).tag(sync=True)
-
-    template: Optional[str] = T.Unicode(
-        None,
-        allow_none=True,
-        help="a nunjucks template to use to calculate node sizes",
-    ).tag(sync=True)
-
-
-@W.register
-class NodeColors(Behavior):
-    _model_name: str = T.Unicode("NodeColorModel").tag(sync=True)
-
-    column_name: str = T.Unicode(
-        None,
-        allow_none=True,
-        help="name of the source column to use for node colors.",
-    ).tag(sync=True)
-
-    template: Optional[str] = T.Unicode(
-        None,
-        allow_none=True,
-        help="a nunjucks template to use to calculate colors",
-    ).tag(sync=True)
 
 
 @W.register
@@ -194,6 +36,8 @@ class LinkSelection(Behavior):
 
 @W.register
 class LinkColors(Behavior):
+    """Customize link colors based on a column or template."""
+
     _model_name: str = T.Unicode("LinkColorModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -211,6 +55,8 @@ class LinkColors(Behavior):
 
 @W.register
 class LinkWidths(Behavior):
+    """Customize link widths based on a column or template."""
+
     _model_name: str = T.Unicode("LinkWidthModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -228,6 +74,8 @@ class LinkWidths(Behavior):
 
 @W.register
 class LinkLabels(Behavior):
+    """Customize link labels, displayed on hover, based on a column or template."""
+
     _model_name: str = T.Unicode("LinkLabelModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -245,6 +93,8 @@ class LinkLabels(Behavior):
 
 @W.register
 class LinkDirectionalArrowColor(Behavior):
+    """Customize arrow color on links, based on a column or template."""
+
     _model_name: str = T.Unicode("LinkDirectionalArrowColorModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -262,6 +112,8 @@ class LinkDirectionalArrowColor(Behavior):
 
 @W.register
 class LinkDirectionalArrowLength(Behavior):
+    """Customize arrow direction on links, based on a column or template."""
+
     _model_name: str = T.Unicode("LinkDirectionalArrowLengthModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -279,6 +131,11 @@ class LinkDirectionalArrowLength(Behavior):
 
 @W.register
 class LinkDirectionalArrowRelPos(Behavior):
+    """Customize arrow positioning on links, based on a column or template.
+
+    This generated value must be between ``0.0``, at the source, or ``1.0``, at the target
+    """
+
     _model_name: str = T.Unicode("LinkDirectionalArrowRelPosModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -296,6 +153,8 @@ class LinkDirectionalArrowRelPos(Behavior):
 
 @W.register
 class LinkDirectionalParticleColor(Behavior):
+    """Customize the color of animated particles on links, based on a column or template."""
+
     _model_name: str = T.Unicode("LinkDirectionalParticleColorModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -313,6 +172,12 @@ class LinkDirectionalParticleColor(Behavior):
 
 @W.register
 class LinkDirectionalParticleSpeed(Behavior):
+    """Customize the speed of animated particles on links, based on a column or template.
+
+    The value should be between ``0.0``, stationary, and ``~0.1``, or they will
+    exceed the frame rate of the animation.
+    """
+
     _model_name: str = T.Unicode("LinkDirectionalParticleSpeedModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -330,6 +195,12 @@ class LinkDirectionalParticleSpeed(Behavior):
 
 @W.register
 class LinkDirectionalParticleWidth(Behavior):
+    """Customize the size of animated particles on links, based on a column or template.
+
+    The value should be between ``0.0``, invisible (not recommended) and ``~5.0``,
+    which is rather large, depending on the size of nodes.
+    """
+
     _model_name: str = T.Unicode("LinkDirectionalParticleWidthModel").tag(sync=True)
 
     column_name: str = T.Unicode(
@@ -347,6 +218,12 @@ class LinkDirectionalParticleWidth(Behavior):
 
 @W.register
 class LinkDirectionalParticles(Behavior):
+    """Customize the density of animated particles on links, based on a column or template.
+
+    The value should be between ``0.0``, invisible (not recommended) and an upper
+    bound, dependant on the visual length of links.
+    """
+
     _model_name: str = T.Unicode("LinkDirectionalParticlesModel").tag(sync=True)
 
     column_name: str = T.Unicode(
