+++ src/ipyforcegraph/sources/__init__.py
@@ -1,2 +1,11 @@
+"""Graph data sources for ``ipyforcegraph``."""
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
+
+from .dataframe import DataFrameSource
+from .widget import WidgetSource
+
+__all__ = [
+    "DataFrameSource",
+    "WidgetSource",
+]
